ingestion/requirements.txt
python-dotenv
pinecone-client 
openai 
tiktoken 
langchain
langchain_openai
langchain-community 
langchain-core
unstructured